Move the cleaning tasks outside the autodetection
