+++ pkg/autodetect/main_test.go
@@ -9,22 +9,15 @@ import (
 	openapi_v2 "github.com/google/gnostic/openapiv2"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/discovery"
-	"k8s.io/client-go/kubernetes/scheme"
 	restclient "k8s.io/client-go/rest"
 	fakeRest "k8s.io/client-go/rest/fake"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	v1 "github.com/jaegertracing/jaeger-operator/apis/v1"
-	"github.com/jaegertracing/jaeger-operator/pkg/inject"
 )
 
 func TestStart(t *testing.T) {
@@ -578,116 +571,6 @@ func TestAuthDelegatorBecomesUnavailable(t *testing.T) {
 	}
 	b.detectClusterRoles(context.Background())
 	assert.False(t, viper.GetBool("auth-delegator-available"))
-}
-
-func TestCleanDeployments(t *testing.T) {
-	for _, tt := range []struct {
-		cap             string // caption for the test
-		watchNamespace  string // the value for WATCH_NAMESPACE
-		jaegerNamespace string // in which namespace the jaeger exists, empty for non existing
-		deleted         bool   // whether the sidecar should have been deleted
-	}{
-		{
-			cap:             "existing-same-namespace",
-			watchNamespace:  "observability",
-			jaegerNamespace: "observability",
-			deleted:         false,
-		},
-		{
-			cap:             "not-existing-same-namespace",
-			watchNamespace:  "observability",
-			jaegerNamespace: "",
-			deleted:         true,
-		},
-		{
-			cap:             "existing-watched-namespace",
-			watchNamespace:  "observability,other-observability",
-			jaegerNamespace: "other-observability",
-			deleted:         false,
-		},
-		{
-			cap:             "existing-non-watched-namespace",
-			watchNamespace:  "observability",
-			jaegerNamespace: "other-observability",
-			deleted:         true,
-		},
-		{
-			cap:             "existing-watching-all-namespaces",
-			watchNamespace:  v1.WatchAllNamespaces,
-			jaegerNamespace: "other-observability",
-			deleted:         false,
-		},
-	} {
-		t.Run(tt.cap, func(t *testing.T) {
-			// prepare the test data
-			viper.Set(v1.ConfigWatchNamespace, tt.watchNamespace)
-			defer viper.Reset()
-
-			jaeger := v1.NewJaeger(types.NamespacedName{
-				Name:      "my-instance",
-				Namespace: "observability", // at first, it exists in the same namespace as the deployment
-			})
-
-			dep := &appsv1.Deployment{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "mydep",
-					Namespace:   "observability",
-					Annotations: map[string]string{inject.Annotation: jaeger.Name},
-				},
-				Spec: appsv1.DeploymentSpec{
-					Template: corev1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{},
-						Spec: corev1.PodSpec{
-							Containers: []corev1.Container{
-								{
-									Name:  "C1",
-									Image: "image1",
-								},
-							},
-						},
-					},
-				},
-			}
-			dep = inject.Sidecar(jaeger, dep)
-
-			// sanity check
-			require.Equal(t, 2, len(dep.Spec.Template.Spec.Containers))
-
-			// prepare the list of existing objects
-			objs := []runtime.Object{dep}
-			if len(tt.jaegerNamespace) > 0 {
-				jaeger.Namespace = tt.jaegerNamespace // now, it exists only in this namespace
-				objs = append(objs, jaeger)
-			}
-
-			// prepare the client
-			s := scheme.Scheme
-			s.AddKnownTypes(v1.GroupVersion, &v1.Jaeger{})
-			s.AddKnownTypes(v1.GroupVersion, &v1.JaegerList{})
-			cl := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-			b := WithClients(cl, &fakeDiscoveryClient{}, cl)
-
-			// test
-			b.cleanDeployments(context.Background())
-
-			// verify
-			persisted := &appsv1.Deployment{}
-			err := cl.Get(context.Background(), types.NamespacedName{
-				Namespace: dep.Namespace,
-				Name:      dep.Name,
-			}, persisted)
-			require.NoError(t, err)
-
-			// should the sidecar have been deleted?
-			if tt.deleted {
-				assert.Equal(t, 1, len(persisted.Spec.Template.Spec.Containers))
-				assert.NotContains(t, persisted.Labels, inject.Label)
-			} else {
-				assert.Equal(t, 2, len(persisted.Spec.Template.Spec.Containers))
-				assert.Contains(t, persisted.Labels, inject.Label)
-			}
-		})
-	}
 }
 
 type fakeClient struct {
